move existing code to base dir
